requirements-dev.txt
# Development and testing dependencies
pytest==8.3.5
pytest-flask==1.3.0
pytest-mock==3.14.0
pytest-cov==6.0.0
freezegun==1.5.1
responses==0.25.3
